Refactor Elector, Reputator and Validator file structure
