+++ nebula/core/SDFL/TrustNode.py
@@ -1,15 +1,12 @@
 import asyncio
 import logging
 
-from nebula.config.config import Config
 from nebula.core.eventmanager import EventManager
 from nebula.core.nebulaevents import RoundEndEvent, RoundStartEvent, ValidationEvent
 from nebula.core.network.communications import CommunicationsManager
-from nebula.core.noderole import AggregatorNode
-from nebula.core.SDFL.elector import Elector, create_elector, get_elector_string, publish_election_event
-from nebula.core.SDFL.reputator import Reputator, create_reputator, get_reputator_string
-from nebula.core.SDFL.validator import Validator, create_validator, get_validator_string
-from nebula.core.training.lightning import Lightning
+from nebula.core.SDFL.Electors.elector import Elector, get_elector_string
+from nebula.core.SDFL.Reputators.reputator import Reputator, get_reputator_string
+from nebula.core.SDFL.Validators.validator import Validator, get_validator_string
 
 
 class TrustNode:
@@ -54,7 +51,7 @@ class TrustNode:
 
     async def _elect_leader(self, re: RoundStartEvent):
         cm: CommunicationsManager = CommunicationsManager.get_instance()
-        self._leader = await self.elector.elect(re)
+        self._leader = await self.elector.elect(re, self.represented_nodes)
         r, _, _ = await re.get_event_data()
         m = cm.create_message("leader", "elect", leader_addr=self.leader, round=r)
         for n in self.represented_nodes:
@@ -83,7 +80,7 @@ class TrustNode:
         if not await self.validator.validate(ve):
             logging.info("Validation failed")
 
-    async def _update_represented(self, new_rep):
+    async def _adjust_represented(self, new_rep):
         cm: CommunicationsManager = CommunicationsManager.get_instance()
         removed_nodes = []
 
@@ -108,7 +105,7 @@ class TrustNode:
         await self._add_trust_node(message.node_addr)
 
     async def _add_trust_node(self, node_addr):
-        rep = await self._update_represented(node_addr)
+        rep = await self._adjust_represented(node_addr)
         if rep:
             cm: CommunicationsManager = CommunicationsManager.get_instance()
             m = cm.create_message(
@@ -127,78 +124,3 @@ class TrustNode:
     async def update_represented(self, rep):
         async with self._lock:
             self.represented_nodes.update(rep)
-
-
-class FollowerNode(AggregatorNode):
-    def __init__(
-        self,
-        model,
-        datamodule,
-        representative,
-        trust_node=None,
-        config=Config,
-        trainer=Lightning,
-        security=False,
-    ):
-        self.leader = None
-        self._lock = asyncio.Lock()
-        self.representative = representative
-        self._trust_node = trust_node
-        super().__init__(
-            model,
-            datamodule,
-            config,
-            trainer,
-            security,
-        )
-
-    @property
-    def trust_node(self):
-        return self._trust_node
-
-    async def start_communications(self):
-        await super().start_communications()
-
-        em: EventManager = EventManager.get_instance()
-        # subscribe to promotion events
-        await em.subscribe(("info", "trust_info"), self._promote_node)
-        # subscribe to representative updates
-        await em.subscribe(("representative", "update"), self._update_representative)
-        # subscribe to leader update
-        await em.subscribe(("leader", "elect"), self._update_leader)
-        # trust node callbacks
-        if self.trust_node is not None:
-            await self._trust_node.start_communications()
-
-    async def _update_representative(self, source, message):
-        if source == self.representative and self.trust_node is None:
-            self.representative = message.node_addr
-
-    async def _promote_node(self, source, message):
-        if not self._trust_node:
-            r = create_reputator(self.config)
-            e = create_elector(self.config, message.represented, message.trusted)
-            v = create_validator(self.config)
-            trusted = [f"{self.ip}:{self.port}"]
-            rep = [f"{self.ip}:{self.port}"]
-            trusted.extend(message.trusted)
-            rep.extend(message.represented)
-
-            self._trust_node = TrustNode(
-                represented_nodes=rep,
-                trusted_nodes=trusted,
-                ip=self.ip,
-                port=self.port,
-                elector=e,
-                validator=v,
-                reputator=r,
-            )
-            self.representative = None
-            await self._trust_node.start_communications()
-        else:
-            await self._trust_node.update_represented(message.represented)
-
-    async def _update_leader(self, source, message):
-        async with self._lock:
-            self.leader = message.leader_addr
-            await publish_election_event(self.leader, message.round)
