+++ nebula/core/SDFL/Validators/NoValidator.py
@@ -1,0 +1,13 @@
+from nebula.core.SDFL.Validators.validator import Validator
+
+
+class NoValidator(Validator):
+    """
+    Implementation to skip validation of proposed model.
+    """
+
+    async def validate(self, model):
+        """
+        Always returns True. Skips validation of proposed model.
+        """
+        return True
